PERF: only register a subset of icons on install
SVG icons are automatically registered in plugins. Update the initializer
to update the core SVG icon set instead to work around this.

Then only register icons that are necessary for the basic swap functions.

This allows all icons to be made available to the instance without all of
the icons registered at once.